+++ tests/test_bundle.py
@@ -22,7 +22,7 @@ import dss
 from dss import DSSException
 from dss.config import BucketConfig, Config, override_bucket_config, Replica
 from dss.util import UrlBuilder
-from dss.util.blobstore import test_object_exists
+from dss.storage.blobstore import test_object_exists
 from dss.util.version import datetime_to_version_format
 from dss.storage.bundles import get_bundle_from_bucket
 from tests.infra import DSSAssertMixin, DSSUploadMixin, ExpectedErrorFields, get_env, testmode
